### Describe your changes

Example of the ~1ms latency between ASGI messages:

```
# Emitting from asgi_app_wrapper
1708486469.33347 msg={'type': 'http.response.start', 'status': 200, 'headers': [(b'content-length', b'14'), (b'content-type', b'application/json')]}    GET / -> 200 OK  (duration: 1.75 s)
1708486469.33355 msg={'type': 'http.response.body', 'body': b'{"name":"foo"}'}
# Emitting from `async for value in res
1708486469.33368 value={'type': 'http.response.start', 'status': 200, 'headers': [(b'content-length', b'14'), (b'content-type', b'application/json')]}
1708486469.33480 value={'type': 'http.response.body', 'body': b'{"name":"foo"}'}
```

☝️ body emitted 1.12ms later, even though it was pushed out of the wrapper only 0.08ms later